backend/requirements.txt
certifi==2019.9.11
chardet==3.0.4
coverage==4.5.4
Django==2.2.6
django-cors-headers==3.1.1
django-filter==2.2.0
djangorestframework==3.10.3
idna==2.8
model-mommy==1.6.0
numpy==1.17.3
pandas==0.25.2
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
